TET-274: remove redundant type files
