--- src/types/Face.ts
@@ -1,7 +1,0 @@
-export type Face = {
-  index: number
-  vertex1: number
-  vertex2: number
-  vertex3: number
-  vertex4: number
-}
