--- src/types/Legend.ts
@@ -1,5 +1,0 @@
-export type LegendType = {
-  rangeStart: number
-  rangeEnd: number
-  color: number[]
-}
