--- src/types/ModelPhysicalQuantity.ts
@@ -1,5 +1,0 @@
-export type ModelPhysicalQuantity = {
-  values: number[]
-  min: number
-  max: number
-}
