--- src/types/Stress.ts
@@ -1,9 +1,0 @@
-export type Stress = {
-  index: number
-  qx: number
-  txy: number
-  tzx: number
-  qy: number
-  tyz: number
-  qz: number
-}
