--- src/types/Vertex.ts
@@ -1,6 +1,0 @@
-export type Vertex = {
-  index: number
-  x: number
-  y: number
-  z: number
-}
